# docs/swagger.yaml
components:
  schemas:
...
    models.ArmorResult:
      properties:
        armor:
          items:
            $ref: '#/components/schemas/models.StrippedItem'
          type: array
          uniqueItems: false
        set_name:
          type: string
        set_rarity:
          type: string
        stats:
          additionalProperties:
            type: number
          type: object
      type: object
...
    models.Commissions:
      properties:
        completions:
          type: integer
        milestone:
          type: integer
      type: object
    models.Contest:
      properties:
        amount:
          type: integer
        collected:
          type: integer
        maxed:
          type: boolean
        medals:
          additionalProperties:
            type: integer
          type: object
        name:
          type: string
        texture:
          type: string
      type: object
    models.Corpse:
      properties:
        amount:
          type: integer
        name:
          type: string
        texture_path:
          type: string
      type: object
    models.Corpses:
      properties:
        corpses:
          items:
            $ref: '#/components/schemas/models.Corpse'
          type: array
          uniqueItems: false
        found:
          type: integer
        max:
          type: integer
      type: object
...
    models.CrystalHollows:
      properties:
        crystalHollowsLastAccess:
          type: integer
        nucleusRuns:
          type: integer
        progress:
          $ref: '#/components/schemas/models.CrystalNucleusRuns'
      type: object
    models.CrystalNucleusRuns:
      properties:
        crystals:
          additionalProperties:
            type: string
          type: object
        parts:
          additionalProperties:
            type: string
          type: object
      type: object
...
    models.EnchantingGame:
      properties:
        attempts:
          type: integer
        bestScore:
          type: integer
        claims:
          type: integer
        name:
          type: string
        texture:
          type: string
      type: object
    models.EnchantingGameData:
      properties:
        name:
          type: string
        stats:
          $ref: '#/components/schemas/models.EnchantingGameStats'
      type: object
    models.EnchantingGameStats:
      properties:
        bonusClicks:
          type: integer
        games:
          items:
            $ref: '#/components/schemas/models.EnchantingGame'
          type: array
          uniqueItems: false
        lastAttempt:
          type: integer
        lastClaimed:
          type: integer
      type: object
    models.EnchantingOutput:
      properties:
        data:
          additionalProperties:
            $ref: '#/components/schemas/models.EnchantingGameData'
          type: object
        unlocked:
          type: boolean
      type: object
    models.EquipmentResult:
      properties:
        equipment:
          items:
            $ref: '#/components/schemas/models.StrippedItem'
          type: array
          uniqueItems: false
        stats:
          additionalProperties:
            type: number
          type: object
      type: object
    models.FairySouls:
      properties:
        found:
          type: integer
        total:
          type: integer
      type: object
    models.FarmingOutput:
      properties:
        contests:
          additionalProperties:
            $ref: '#/components/schemas/models.Contest'
          type: object
        contestsAttended:
          type: integer
        copper:
          type: integer
        medals:
          additionalProperties:
            $ref: '#/components/schemas/models.Medal'
          type: object
        pelts:
          type: integer
        tools:
          $ref: '#/components/schemas/models.SkillToolsResult'
        uniqueGolds:
          type: integer
      type: object
    models.FishingOuput:
      properties:
        itemsFished:
          type: integer
        kills:
          items:
            $ref: '#/components/schemas/models.Kill'
          type: array
          uniqueItems: false
        seaCreaturesFished:
          type: integer
        shredderBait:
          type: integer
        shredderFished:
          type: integer
        tools:
          $ref: '#/components/schemas/models.SkillToolsResult'
        treasure:
          type: integer
        treasureLarge:
          type: integer
        trophyFish:
          $ref: '#/components/schemas/models.TrophyFishOutput'
      type: object
    models.ForgeOutput:
      properties:
        duration:
          type: number
        endingTime:
          type: integer
        id:
          type: string
        name:
          type: string
        slot:
          type: integer
        startingTime:
          type: integer
      type: object
...
    models.Fossil:
      properties:
        found:
          type: boolean
        name:
          type: string
        texture_path:
          type: string
      type: object
    models.Fossils:
      properties:
        fossils:
          items:
            $ref: '#/components/schemas/models.Fossil'
          type: array
          uniqueItems: false
        found:
          type: integer
        max:
          type: integer
      type: object
...
    models.GlaciteTunnels:
      properties:
        corpses:
          $ref: '#/components/schemas/models.Corpses'
        fossilDust:
          type: number
        fossils:
          $ref: '#/components/schemas/models.Fossils'
        mineshaftsEntered:
          type: integer
      type: object
    models.HotmTokens:
      properties:
        available:
          type: integer
        spent:
          type: integer
        total:
          type: integer
      type: object
    models.Kill:
      properties:
        amount:
          type: integer
        id:
          type: string
        name:
          type: string
        texture:
          type: string
      type: object
    models.Medal:
      properties:
        amount:
          type: integer
        total:
          type: integer
      type: object
    models.MemberStats:
      properties:
        cute_name:
          type: string
        profile_id:
          type: string
        removed:
          type: boolean
        username:
          type: string
        uuid:
          type: string
      type: object
    models.MiningOutput:
      properties:
        commissions:
          $ref: '#/components/schemas/models.Commissions'
        crystalHollows:
          $ref: '#/components/schemas/models.CrystalHollows'
        forge:
          items:
            $ref: '#/components/schemas/models.ForgeOutput'
          type: array
          uniqueItems: false
        glaciteTunnels:
          $ref: '#/components/schemas/models.GlaciteTunnels'
        hotm:
          items:
            $ref: '#/components/schemas/models.ProcessedItem'
          type: array
          uniqueItems: false
        level:
          $ref: '#/components/schemas/models.Skill'
        peakOfTheMountain:
          $ref: '#/components/schemas/models.PeakOfTheMountain'
        powder:
          $ref: '#/components/schemas/models.PowderOutput'
        selectedPickaxeAbility:
          type: string
        tokens:
          $ref: '#/components/schemas/models.HotmTokens'
        tools:
          $ref: '#/components/schemas/models.SkillToolsResult'
      type: object
...
    models.PeakOfTheMountain:
      properties:
        level:
          type: integer
        max_level:
          type: integer
      type: object
...
    models.PowderAmount:
      properties:
        available:
          type: integer
        spent:
          type: integer
        total:
          type: integer
      type: object
    models.PowderOutput:
      properties:
        gemstone:
          $ref: '#/components/schemas/models.PowderAmount'
        glacite:
          $ref: '#/components/schemas/models.PowderAmount'
        mithril:
          $ref: '#/components/schemas/models.PowderAmount'
      type: object
    models.ProcessedItem:
      properties:
        Count:
          type: integer
        Damage:
          type: integer
        categories:
          items:
            type: string
          type: array
          uniqueItems: false
        containsItems:
          items:
            $ref: '#/components/schemas/models.ProcessedItem'
          type: array
          uniqueItems: false
        display_name:
          type: string
        id:
          type: string
        isInactive:
          type: boolean
        lore:
          items:
            type: string
          type: array
          uniqueItems: false
        price:
          type: number
        rarity:
          type: string
        recombobulated:
          type: boolean
        shiny:
          type: boolean
        source:
          type: string
        tag:
          $ref: '#/components/schemas/skycrypttypes.Tag'
        texture_pack:
          type: string
        texture_path:
          type: string
        wiki:
          $ref: '#/components/schemas/models.WikipediaLinks'
      type: object
    models.ProcessingError:
      properties:
        error:
          type: string
        message:
          type: string
        status:
          type: string
      type: object
    models.ProfilesStats:
      properties:
        cute_name:
          type: string
        game_mode:
          type: string
        profile_id:
          type: string
        selected:
          type: boolean
      type: object
    models.RankOutput:
      properties:
        plusColor:
          type: string
        plusText:
          type: string
        rankColor:
          type: string
        rankText:
          type: string
      type: object
...
    models.RiftCastleOutput:
      properties:
        grubberStacks:
          type: integer
        maxBurgers:
          type: integer
      type: object
    models.RiftEnigmaOutput:
      properties:
        souls:
          type: integer
        totalSouls:
          type: integer
      type: object
    models.RiftMotesOutput:
      properties:
        lifetime:
          type: integer
        orbs:
          type: integer
        purse:
          type: integer
      type: object
    models.RiftOutput:
      properties:
        armor:
          $ref: '#/components/schemas/models.ArmorResult'
        castle:
          $ref: '#/components/schemas/models.RiftCastleOutput'
        enigma:
          $ref: '#/components/schemas/models.RiftEnigmaOutput'
        equipment:
          $ref: '#/components/schemas/models.EquipmentResult'
        motes:
          $ref: '#/components/schemas/models.RiftMotesOutput'
        porhtal:
          $ref: '#/components/schemas/models.RiftPortalsOutput'
        timecharms:
          $ref: '#/components/schemas/models.RiftTimecharmsOutput'
        visits:
          type: integer
      type: object
    models.RiftPorhtal:
      properties:
        name:
          type: string
        texture:
          type: string
        unlocked:
          type: boolean
      type: object
    models.RiftPortalsOutput:
      properties:
        porhtals:
          items:
            $ref: '#/components/schemas/models.RiftPorhtal'
          type: array
          uniqueItems: false
        porhtalsFound:
          type: integer
      type: object
    models.RiftTimecharms:
      properties:
        name:
          type: string
        texture:
          type: string
        unlocked:
          type: boolean
        unlockedAt:
          type: integer
      type: object
    models.RiftTimecharmsOutput:
      properties:
        timecharms:
          items:
            $ref: '#/components/schemas/models.RiftTimecharms'
          type: array
          uniqueItems: false
        timecharmsFound:
          type: integer
      type: object
...
    models.Skill:
      properties:
        level:
          type: integer
        levelCap:
          type: integer
        levelWithProgress:
          type: number
        maxLevel:
          type: integer
        maxed:
          type: boolean
        progress:
          type: number
        texture:
          type: string
        uncappedLevel:
          type: integer
        unlockableLevelWithProgress:
          type: number
        xp:
          type: integer
        xpCurrent:
          type: integer
        xpForNext:
          type: integer
      type: object
    models.SkillToolsResult:
      properties:
        highest_priority_tool:
          $ref: '#/components/schemas/models.StrippedItem'
        tools:
          items:
            $ref: '#/components/schemas/models.StrippedItem'
          type: array
          uniqueItems: false
      type: object
    models.Skills:
      properties:
        averageSkillLevel:
          type: number
        averageSkillLevelWithProgress:
          type: number
        skills:
          additionalProperties:
            $ref: '#/components/schemas/models.Skill'
          type: object
        totalSkillXp:
          type: integer
      type: object
    models.SkillsOutput:
      properties:
        enchanting:
          $ref: '#/components/schemas/models.EnchantingOutput'
        farming:
          $ref: '#/components/schemas/models.FarmingOutput'
        fishing:
          $ref: '#/components/schemas/models.FishingOuput'
        mining:
          $ref: '#/components/schemas/models.MiningOutput'
      type: object
...
    models.SourceTab:
      properties:
        icon:
          type: string
        name:
          type: string
      type: object
...
    models.StatsOutput:
      properties:
        apiSettings:
          additionalProperties:
            type: boolean
          type: object
        bank:
          type: number
        displayName:
          type: string
        fairySouls:
          $ref: '#/components/schemas/models.FairySouls'
        game_mode:
          type: string
        joined:
          type: integer
        members:
          items:
            $ref: '#/components/schemas/models.MemberStats'
          type: array
          uniqueItems: false
        personalBank:
          type: number
        profile_cute_name:
          type: string
        profile_id:
          type: string
        profiles:
          items:
            $ref: '#/components/schemas/models.ProfilesStats'
          type: array
          uniqueItems: false
        purse:
          type: number
        rank:
          $ref: '#/components/schemas/models.RankOutput'
        selected:
          type: boolean
        skills:
          $ref: '#/components/schemas/models.Skills'
        skyblock_level:
          $ref: '#/components/schemas/models.Skill'
        social:
          $ref: '#/components/schemas/skycrypttypes.SocialMediaLinks'
        username:
          type: string
        uuid:
          type: string
      type: object
    models.StrippedItem:
      properties:
        Count:
          type: integer
        containsItems:
          items:
            $ref: '#/components/schemas/models.StrippedItem'
          type: array
          uniqueItems: false
        display_name:
          type: string
        isInactive:
          type: boolean
        lore:
          items:
            type: string
          type: array
          uniqueItems: false
        rarity:
          type: string
        recombobulated:
          type: boolean
        shiny:
          type: boolean
        source:
          type: string
        sourceTab:
          $ref: '#/components/schemas/models.SourceTab'
        texture_pack:
          type: string
        texture_path:
          type: string
        wiki:
          $ref: '#/components/schemas/models.WikipediaLinks'
      type: object
...
    models.TrophyFish:
      properties:
        bronze:
          type: integer
        description:
          type: string
        diamond:
          type: integer
        gold:
          type: integer
        id:
          type: string
        maxed:
          type: boolean
        name:
          type: string
        silver:
          type: integer
        texture:
          type: string
      type: object
    models.TrophyFishOutput:
      properties:
        stage:
          $ref: '#/components/schemas/models.TrophyFishStage'
        totalCaught:
          type: integer
        trophyFish:
          items:
            $ref: '#/components/schemas/models.TrophyFish'
          type: array
          uniqueItems: false
      type: object
    models.TrophyFishProgress:
      properties:
        caught:
          type: integer
        tier:
          type: string
        total:
          type: integer
      type: object
    models.TrophyFishStage:
      properties:
        name:
          type: string
        progress:
          items:
            $ref: '#/components/schemas/models.TrophyFishProgress'
          type: array
          uniqueItems: false
      type: object
...
    models.WikipediaLinks:
      properties:
        fandom:
          type: string
        official:
          type: string
      type: object
...
    skycrypttypes.Display:
      properties:
        Lore:
          items:
            type: string
          type: array
          uniqueItems: false
        Name:
          type: string
        color:
          type: integer
      type: object
    skycrypttypes.ExtraAttributes:
      description: |-
        HideFlags       int             `nbt:"HideFlags" json:"HideFlags,omitempty"`
        Unbreakable     int             `nbt:"Unbreakable" json:"Unbreakable,omitempty"`
        Enchantments    []Enchantment   `nbt:"ench" json:"ench,omitempty"`
      properties:
        ability_scroll:
          items:
            type: string
          type: array
          uniqueItems: false
        additional_coins:
          type: integer
        art_of_war_count:
          type: integer
        artOfPeaceApplied:
          type: integer
        attributes:
          additionalProperties:
            type: integer
          type: object
        auction:
          type: integer
        bid:
          type: integer
        boosters:
          items:
            type: string
          type: array
          uniqueItems: false
        champion_combat_xp:
          type: number
        collected_coins:
          type: integer
        compact_blocks:
          type: integer
        divan_powder_coating:
          type: integer
        donated_museum:
          type: boolean
        drill_part_engine:
          type: string
        drill_part_fuel_tank:
          type: string
        drill_part_upgrade_module:
          type: string
        dungeon_item_level: {}
        dye_item:
          type: string
        edition:
          type: integer
        enchantments:
          additionalProperties:
            type: integer
          type: object
        ethermerge:
          type: integer
        expertise_kills:
          type: integer
        farmed_cultivating:
          type: integer
        farming_for_dummies_count:
          type: integer
        gems:
          additionalProperties: {}
          type: object
        hecatomb_s_runs:
          type: integer
        hook:
          $ref: '#/components/schemas/skycrypttypes.RodPart'
        hot_potato_count:
          type: integer
        id:
          type: string
        is_shiny:
          type: boolean
        item_tier:
          type: integer
        jalapeno_count:
          type: integer
        levelable_overclocks:
          type: integer
        line:
          $ref: '#/components/schemas/skycrypttypes.RodPart'
        mana_disintegrator_count:
          type: integer
        model:
          type: string
        modifier:
          type: string
        new_year_cake_bag_data:
          items:
            type: integer
          type: array
          uniqueItems: false
        new_year_cake_bag_years:
          items:
            type: integer
          type: array
          uniqueItems: false
        new_years_cake:
          type: integer
        party_hat_color:
          type: string
        party_hat_emoji:
          type: string
        petInfo:
          type: string
        pickonimbus_durability:
          type: integer
        polarvoid:
          type: integer
        power_ability_scroll:
          type: string
        price:
          type: integer
        rarity_upgrades:
          type: integer
        runes:
          additionalProperties:
            type: integer
          type: object
        sack_pss:
          type: integer
        sinker:
          $ref: '#/components/schemas/skycrypttypes.RodPart'
        skin:
          type: string
        talisman_enrichment:
          type: string
        thunder_charge:
          type: integer
        timestamp: {}
        tuned_transmission:
          type: integer
        upgrade_level: {}
        uuid:
          type: string
        winning_bid:
          type: integer
        wood_singularity_count:
          type: integer
      type: object
...
    skycrypttypes.Properties:
      properties:
        textures:
          items:
            $ref: '#/components/schemas/skycrypttypes.Texture'
          type: array
          uniqueItems: false
      type: object
    skycrypttypes.RodPart:
      properties:
        donated_museum:
          type: boolean
        part:
          type: string
      type: object
    skycrypttypes.SkullOwner:
      properties:
        Id:
          type: string
        Properties:
          $ref: '#/components/schemas/skycrypttypes.Properties'
      type: object
    skycrypttypes.SocialMediaLinks:
      properties:
        DISCORD:
          type: string
        HYPIXEL:
          type: string
        TWITCH:
          type: string
        TWITTER:
          type: string
      type: object
    skycrypttypes.Tag:
      properties:
        ExtraAttributes:
          $ref: '#/components/schemas/skycrypttypes.ExtraAttributes'
        SkullOwner:
          $ref: '#/components/schemas/skycrypttypes.SkullOwner'
        display:
          $ref: '#/components/schemas/skycrypttypes.Display'
      type: object
    skycrypttypes.Texture:
      properties:
        Signature:
          type: string
        Value:
          type: string
      type: object
...
info:
  description: API for SkyCrypt - A Hypixel SkyBlock Stats Viewer
  title: SkyCrypt API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /api/rift/{uuid}/{profileId}:
    get:
      description: Returns rift data for the given user and profile ID
      parameters:
      - description: User UUID
        in: path
        name: uuid
        required: true
        schema:
          type: string
      - description: Profile ID
        in: path
        name: profileId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.RiftOutput'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ProcessingError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ProcessingError'
          description: Internal Server Error
      summary: Get rift stats of a specified player
      tags:
      - rift
  /api/skills/{uuid}/{profileId}:
    get:
      description: Returns skills for the given user and profile ID
      parameters:
      - description: User UUID
        in: path
        name: uuid
        required: true
        schema:
          type: string
      - description: Profile ID
        in: path
        name: profileId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.SkillsOutput'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ProcessingError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ProcessingError'
          description: Internal Server Error
      summary: Get skills stats of a specified player
      tags:
      - skills
...
  /api/stats/{uuid}/{profileId}:
    get:
      description: Returns stats for the given user and profile ID
      parameters:
      - description: User UUID
        in: path
        name: uuid
        required: true
        schema:
          type: string
      - description: Profile ID
        in: path
        name: profileId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.StatsOutput'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ProcessingError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ProcessingError'
          description: Internal Server Error
      summary: Get stats of a specified player
      tags:
      - stats
...
servers:
- url: localhost:8080/